Add a signal handler to clean up after ctrl-c
Add signal handers into platform.cpp. A process can register a single signal
handler and the last registered handler is the one that takes effect. The
handler in OpenCL is positioned to override the IPC signal handler set up in
Ipc_start() on AM57.

If the application sets up a signal handler, it will override the one in
OpenCL, as expected.