OpenCL RTOS: Check layout of DSP exported syms
- In OpenCL RTOS, DSP exported syms need to be laid out exactly the same
on both cores, so that when OpenCL kernel is compiled and linked against
symbols on core 0, it will be also correct for core 1 during runtime.
Added a function in the Makefile to check for the layout.  User can enable
calling the function to check the dsp executables for core 0 and core 1.