MCT-608: Prevent exit/abort from C code impacting monitor execution
Calls to abort, exit or assert from OpenCL C code or C code called from
OpenCL C kernels should not terminate the montor.

abort and exit are replaced by corresponding versions in the monitor
that report to the exit/abort to the user and return to normal monitor
execution i.e. listen to messages from the host.