Add option to specify size of shared memory allocated by ti-mctd
OpenCL maintains a heap shared across Linux processes. The heap is in
CMEM, but the metadata (alloc, free lists) is in Linux shared memory
allocated by ti-mctd. It is possible for an application to run out of
shared memory if it allocates a lot of small buffers or if there is
fragmentation.

Adding an option enables the user to increase the size of shared memory.

(MCT-702)