Fix kernel error message at AM5749 IDK boot
- The generic EVE firmware contains 4 EVE images, which can be used
for both AM572x and AM574x.  EVEs have been conditionally booted up
according to the number of EVEs available on the device.
However, we missed one place that loads the EVE image in SBL lib,
which parses the firmware and loads all available images.
- The fix is in IPU, before handing eve firmware to SBL lib image
parser, to mark extra EVEs as unavailable, so that SBL lib image parser
will only load to the EVEs available on the device.
- MCT-993